Revert "refactor: 주식 상세화면 폴더 구조 변경"
