+++ src/pages/stock-detail/ui/comment-form.tsx
