+++ src/pages/stock-detail/ui/stock-analysis-table.tsx
