+++ src/pages/stock-detail/ui/stock-detail-page.tsx
@@ -1,10 +1,9 @@
 import { useParams } from 'react-router';
 import { useQueryState } from 'nuqs';
 import { Tab, Tabs } from '@/shared/ui';
-import { StockAnalyst } from './stock-analyst/stock-analyst';
-
+import { StockAnalyst } from './stock-analyst';
+import { Community } from './community';
 import { StockHeader } from './stock-header';
-import { Community } from './community/community';
 
 export function StockDetailPage() {
   const { stockCode } = useParams<{ stockCode: string }>();
